configure.ac: change SDL2 check to work on mingw-w64
One only needs -lSDL2main and -mwindows when building an application
linked against SDL2, not a library which is linked against SDL2.
Without this change you'll only be able to build a static libSDL_sound.a
because libtool freaks over libmingw32.a and libSDL2main.a not having a
real library/dll to go with them.

Signed-off-by: Marty E. Plummer <hanetzer@protonmail.com>